# Conflicts:
#   app/Config/Routes.php
#   app/Controllers/Admin/DevelopmentTypesController.php
#   app/Views/pages/admin/alerts.php

+++ app/Controllers/Admin/CompetitionTypeController.php
@@ -13,7 +13,7 @@ class CompetitionTypeController extends BaseController
 
         $data = [
             'competitionType' => $competitionType->findAll(),
-            'title'           => 'Competitions Type',
+            'title'           => 'Competition Types',
         ];
 
         return view('pages/admin/competitionType', $data);
@@ -72,6 +72,12 @@ class CompetitionTypeController extends BaseController
     public function delete($id = null)
     {
         $competitionType = new competitionTypeModel();
+
+
+        if(sizeof($competitionType->findCompetitions($id)) != 0) {
+            return redirect()->back()->with('alert', ['type' => 'danger', 'content' => "Make sure all of the developments associated with this type are deleted first."]);
+        }
+
         $competitionType->delete($id);
 
         return redirect()->to(base_url('admin/CompetitionType'))->with('status', 'successes');
